// admin/src/modulos/api/model/paginationEspecialidade.ts
/**
 * Receita Digital
 * Definição da API da Receita Digital 
 *
 * OpenAPI spec version: 1.0.6
 * Contact: fran.fig@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Especialidade } from './especialidade';


export interface PaginationEspecialidade {
    total?: number;
    limit?: number;
    page?: number;
    docs?: Array<Especialidade>;
}